src/test/java/de/adorsys/keycloak/config/properties/ImportConfigPropertiesTest.java
/*
 * Copyright 2019-2020 adorsys GmbH & Co. KG @ https://adorsys.de
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 */

package de.adorsys.keycloak.config.properties;

import de.adorsys.keycloak.config.properties.ImportConfigProperties.ImportManagedProperties.ImportManagedPropertiesValues;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.junit.jupiter.SpringExtension;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.core.Is.is;

// From: https://tuhrig.de/testing-configurationproperties-in-spring-boot/
@ExtendWith(SpringExtension.class)
@SpringBootTest(classes = {ImportConfigPropertiesTest.TestConfiguration.class})
@TestPropertySource(properties = {
        "spring.main.log-startup-info=false",
        "import.cache-key=custom",
        "import.force=true",
        "import.path=other",
        "import.managed.group=no-delete",
        "import.managed.required-action=no-delete",
        "import.managed.client-scope=no-delete",
        "import.managed.scope-mapping=no-delete",
})
public class ImportConfigPropertiesTest {

    @Autowired
    private ImportConfigProperties properties;

    @Test
    public void shouldPopulateConfigurationProperties() {
        assertThat(properties.getPath(), is("other"));
        assertThat(properties.isForce(), is(true));
        assertThat(properties.getCacheKey(), is("custom"));
        assertThat(properties.getManaged().getGroup(), is(ImportManagedPropertiesValues.noDelete));
        assertThat(properties.getManaged().getRequiredAction(), is(ImportManagedPropertiesValues.noDelete));
        assertThat(properties.getManaged().getClientScope(), is(ImportManagedPropertiesValues.noDelete));
        assertThat(properties.getManaged().getScopeMapping(), is(ImportManagedPropertiesValues.noDelete));
    }

    @EnableConfigurationProperties(ImportConfigProperties.class)
    public static class TestConfiguration {
        // nothing
    }
}
